requirements.txt
sqlalchemy
openai
arxiv
pymupdf
requests
resend
streamlit
python-dotenv
pandas
plotly
backoff